pyproject.toml
[tool.poetry]
name = "mcqgenerator"
version = "0.1.0"
description = ""
authors = ["fahadsalamatrandhawa <dreamtwister61814@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.33.0"
langchain = "^0.1.14"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
google-generativeai = "^0.4.1"
langchain-google-genai = "^1.0.1"
ipykernel = "^6.29.4"
sentence-transformers = "^2.6.1"
langchain-mongodb = "^0.1.3"
pymongo = "^4.6.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
